#feat 사용자 회원가입 기능 추가(진행중)
1) 사용자 회원가입 구현하였습니다.(관리자는 아직입니다.)
2) 회원가입 과정을 따로 처리하는 registerProcess를 만들었습니다.
