+++ myapp/package.json
@@ -13,6 +13,7 @@
     "express": "~4.16.1",
     "express-session": "^1.17.2",
     "http-errors": "~1.6.3",
+    "mariadb": "^2.5.4",
     "morgan": "~1.9.1",
     "mysql": "^2.18.1",
     "nodemailer": "^6.6.2",
