+++ myapp/views/register_for_admin.ejs
@@ -12,7 +12,7 @@
     <div class="Container">
       <!-- login form -->
         <div class="login-form">
-          <form id="submitForm" method="POST" action="/" onsubmit="return checkForm()">
+          <form id="submitForm" method="POST" action="/registerProcessForAdmin" onsubmit="return checkForm()">
             <%-include('./layout/register_form')%>
             <label class="form-label">소속 분류</label>
             <label class="form-text">선택 입니다.</label>
