+++ myapp/views/register_for_user.ejs
@@ -11,12 +11,12 @@
     <div class="Container">
       <!-- login form -->
       <div class="login-form">
-        <form id="submitForm" method="POST" action="/" onsubmit="return checkForm()">
+        <form id="submitForm" method="POST" action="/registerProcessForUser" onsubmit="return checkForm()">
           <%-include('./layout/register_form')%>
             <div class="mb-3">
               <label for="registerTeam" class="form-label">소속</label>
               <label class="form-text">학과 또는 동아리</label>
-              <input type="text" class="form-control" id="registerTeam" aria-describedby="emailHelp">
+              <input type="text" class="form-control" name="team" id="registerTeam" aria-describedby="emailHelp">
             </div>
             <button type="submit" class="btn btn-color-orange">가입하기</button>
         </form>
